aimodule/pyproject.toml
[project]
name = "aimodule"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.12.0",
    "google-genai>=1.45.0",
    "langchain>=1.0.0",
    "langchain-community>=0.4",
    "langchain-google-genai>=3.0.0",
    "langgraph>=1.0.0",
    "psycopg2-binary>=2.9.11",
    "pytest>=8.4.2",
]
